Do not hash files in the 'deps' pipeline
The problem is that in the 'deps' pipeline, our `dep.file` might be a
relative path from a directory other than the `cwd`. We could use
`dep.basedir` to resolve the path to an absolute one, but we don't
really need to be hashing the dependencies here, since all files read by
browserify call fire the `file` event with an absolute path.

Instead we follow suit with `watchify` and only set the cache for
browserify in `deps`, and watch files (hash them, record that we depend
on them, etc) in the `file` and `package` events.